Added Polices

+++ app/Http/Controllers/API/LessonController.php
@@ -69,6 +69,9 @@ class LessonController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $idlesson = Lesson::findOrFail($id);
+        $this->authorize('update', $idlesson);
+
         $lesson =new LessonResource(Lesson::findOrFail($id));
         $lesson->update($request->all());
 
@@ -84,6 +87,9 @@ class LessonController extends Controller
      */
     public function destroy($id)
     {
+        $idlesson = Lesson::findOrFail($id);
+        $this->authorize('delete', $idlesson);
+
         Lesson::findOrFail($id)->delete();
         return 204;
     }
